requirements.txt
fastapi~=0.135.2
pydantic~=2.12.5
numpy~=2.4.3
openai~=2.30.0
python-dotenv~=1.2.2
SQLAlchemy~=2.0.48